added a remove button and renamed the imports
