+++ imports/ui/App.jsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
-import Updates from './Updates.jsx'
+import Members from './members.jsx'
 import CreatePost from './dashboard/CreatePost'
 import CreateLeader from './dashboard/leads/CreateLeader'
 import AddNumber from './dashboard/AddNumber'
@@ -29,7 +29,7 @@ const App = () => (
       <Header />
       <br />
         <Route exact path='/' component={Home} />
-        <Route exact path="/updates" component={Updates} />
+        <Route exact path="/users" component={Members} />
         <Route path="/post" component={CreatePost} />
         <Route path="/posts" component={Post} />
         <Route path="/upload" component={CreateLeader} />
